squash compiler warnings

+++ modified-lar/src/main/scala/hmda/publication/lar/parser/ModifiedLarCsvParser.scala
@@ -240,7 +240,7 @@ object ModifiedLarCsvParser {
     if (code == 0) "" else code.toString
 
   private def roundToMidPoint(x: Int): Int = {
-    val rounded = 10000 * Math.floor(x / 10000) + 5000
+    val rounded = 10000 * Math.floor(x.toDouble / 10000) + 5000
     rounded.toDouble.toInt
   }
